app/azure_functions/postgres_ingestor/requirements.txt
azure-functions
azure-storage-blob
azure-identity
psycopg2-binary
pandas
pyarrow

